Renamed StartNewGameOfLife to GameManager and moved method NextGenerationField into it. Deleted PrintEveryArrayIteration.
