+++ GameOfLife/Calculations/CellPopulator.cs
@@ -19,7 +19,7 @@
                 for (int j = 0; j < gameField.GetLength(1); j++)
                 {
                    gameField[i, j] = randomizer.Next(0, 2); 
-                   //GliderCreation(gameField);
+                  // GliderCreation(gameField);
 
                 }
