+++ GameOfLife/Displaying/DisplayGameField.cs
@@ -30,15 +30,5 @@ namespace GameOfLife
             Thread.Sleep(1000);
 
         }
-
-        /// <summary>
-        /// Prints calculated and formated next generation array iteration based on previous 
-        /// </summary>
-        /// <param name="gameField">Array containing previous generation cell values</param>
-        public void PrintEveryArrayIteration(int[,] gameField)
-        {
-            NextCellGeneration nextCellGeneration = new NextCellGeneration();
-            nextCellGeneration.NextGenerationGameField(gameField);
-        }
     }
 }
